extract geo-types into separate module
